src/notepadequalequal/requirements.txt
pyinstaller
pillow
tk
tklinenums
pyperclip
pyspellchecker